authors_service/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
redis
uvicorn
